+++ x/gov/keeper/msg_server.go
@@ -352,6 +352,53 @@ func (k msgServer) UpdateMessageParams(ctx context.Context, msg *v1.MsgUpdateMes
 	return &v1.MsgUpdateMessageParamsResponse{}, nil
 }
 
+// SudoExec implements the v1.MsgServer method
+func (k msgServer) SudoExec(ctx context.Context, msg *v1.MsgSudoExec) (*v1.MsgSudoExecResponse, error) {
+	if msg == nil || msg.Msg == nil {
+		return nil, errors.Wrap(govtypes.ErrInvalidProposal, "sudo-ed message cannot be nil")
+	}
+
+	if k.authority != msg.Authority {
+		return nil, errors.Wrapf(govtypes.ErrInvalidSigner, "invalid authority; expected %s, got %s", k.authority, msg.Authority)
+	}
+
+	sudoedMsg, err := msg.GetSudoedMsg()
+	if err != nil {
+		return nil, errors.Wrapf(govtypes.ErrInvalidProposal, "invalid sudo-ed message: %s", err)
+	}
+
+	// check if the message implements the HasValidateBasic interface
+	if m, ok := sudoedMsg.(sdk.HasValidateBasic); ok {
+		if err := m.ValidateBasic(); err != nil {
+			return nil, errors.Wrapf(govtypes.ErrInvalidProposal, "invalid sudo-ed message: %s", err)
+		}
+	}
+
+	handler := k.router.Handler(sudoedMsg)
+	if handler == nil {
+		return nil, errors.Wrapf(govtypes.ErrInvalidProposal, "unrecognized message route: %s", sdk.MsgTypeURL(sudoedMsg))
+	}
+
+	sdkCtx := sdk.UnwrapSDKContext(ctx)
+	msgResp, err := handler(sdkCtx, sudoedMsg)
+	if err != nil {
+		return nil, errors.Wrapf(err, "failed to execute message; message %v", msg)
+	}
+
+	// emit the events from the executed message
+	events := msgResp.Events
+	sdkEvents := make([]sdk.Event, 0, len(events))
+	for _, event := range events {
+		e := event
+		sdkEvents = append(sdkEvents, sdk.Event(e))
+	}
+	sdkCtx.EventManager().EmitEvents(sdkEvents)
+
+	return &v1.MsgSudoExecResponse{
+		Result: msgResp.Data,
+	}, nil
+}
+
 type legacyMsgServer struct {
 	govAcct string
 	server  v1.MsgServer